+++ app/src/frontend/building/data-containers/land-use.tsx
@@ -13,6 +13,7 @@ import { CategoryViewProps } from './category-view-props';
 import Verification from '../data-components/verification';
 import { useDisplayPreferences } from '../../displayPreferences-context';
 import { DataEntryGroup } from '../data-components/data-entry-group';
+import { t } from 'i18next';
 
 /**
  * Use view/edit section
@@ -35,21 +36,21 @@ const UseView: React.FunctionComponent<CategoryViewProps> = (props) => {
                 <DataEntryGroup name="General Land Use">
                     <div className={`alert alert-dark`} role="alert" style={{ fontSize: 13, backgroundColor: "#f6f8f9" }}>
                         <i>
-                            The vast majority of properties are residential (93% in the UK), so we have set 'residential' as the default value. Can you help us identify non-residential and mixed use buildings (and verify residential buildings too)?
+                            {t("The vast majority of properties are residential (93% in the UK), so we have set 'residential' as the default value. Can you help us identify non-residential and mixed use buildings (and verify residential buildings too)?")}
                         </i>
                     </div>
                     <button className={`map-switcher-inline ${props.mapColourScale == "is_domestic" ? "enabled-state" : "disabled-state"} btn btn-outline btn-outline-dark ${darkLightTheme}`} onClick={switchToIsDomesticMapStyle}>
-                        {(props.mapColourScale == "is_domestic")? 'Showing residential, non-residential and mixed-use buildings (click to hide)' : 'Click to see residential, non-residential and mixed-use buildings on the map.'}
+                        {(props.mapColourScale == "is_domestic")? t('Showing residential, non-residential and mixed-use buildings (click to hide)') : t('Click to see residential, non-residential and mixed-use buildings on the map.')}
                     </button>
                     <SelectDataEntry
-                        title={dataFields.is_domestic.title}
+                        title={t(dataFields.is_domestic.title)}
                         slug="is_domestic"
                         value={props.building.is_domestic}
                         options={dataFields.is_domestic.items}
                         mode={props.mode}
                         copy={props.copy}
                         onChange={props.onChange}
-                        tooltip={dataFields.is_domestic.tooltip}
+                        tooltip={t(dataFields.is_domestic.tooltip)}
                     />
                     <Verification
                         slug="is_domestic"
@@ -60,26 +61,26 @@ const UseView: React.FunctionComponent<CategoryViewProps> = (props) => {
                         verified_count={props.building.verified.is_domestic}
                     />
                     <SelectDataEntry
-                        title={dataFields.is_domestic_source.title}
+                        title={t(dataFields.is_domestic_source.title)}
                         slug="is_domestic_source"
                         value={props.building.is_domestic_source}
                         options={dataFields.is_domestic_source.items}
                         mode={props.mode}
                         copy={props.copy}
                         onChange={props.onChange}
-                        tooltip={dataFields.is_domestic_source.tooltip}
+                        tooltip={t(dataFields.is_domestic_source.tooltip)}
                     />
                     {(props.building.is_domestic_source == commonSourceTypes[0] ||
                         props.building.is_domestic_source == commonSourceTypes[1] ||
                         props.building.is_domestic_source == null) ? <></> :
                         <><MultiDataEntry
-                            title={dataFields.is_domestic_links.title}
+                            title={t(dataFields.is_domestic_links.title)}
                             slug="is_domestic_links"
                             value={props.building.is_domestic_links}
                             mode={props.mode}
                             copy={props.copy}
                             onChange={props.onChange}
-                            tooltip={dataFields.is_domestic_links.tooltip}
+                            tooltip={t(dataFields.is_domestic_links.tooltip)}
                             placeholder="https://..."
                             editableEntries={true}
                             isUrl={true}
@@ -89,14 +90,14 @@ const UseView: React.FunctionComponent<CategoryViewProps> = (props) => {
                 </DataEntryGroup>
                 <DataEntryGroup name="Specific land use(s)">
                     <MultiDataEntry
-                        title={dataFields.current_landuse_group.title}
+                        title={t(dataFields.current_landuse_group.title)}
                         slug="current_landuse_group"
                         value={props.building.current_landuse_group}
                         mode={props.mode}
                         copy={props.copy}
                         onChange={props.onChange}
                         confirmOnEnter={true}
-                        tooltip={dataFields.current_landuse_group.tooltip}
+                        tooltip={t(dataFields.current_landuse_group.tooltip)}
                         placeholder="Type new land use group here"
                         copyable={true}
                         autofill={true}
@@ -111,27 +112,27 @@ const UseView: React.FunctionComponent<CategoryViewProps> = (props) => {
                         verified_count={props.building.verified.current_landuse_group}
                         />
                     <SelectDataEntry
-                        title={dataFields.current_landuse_source.title}
+                        title={t(dataFields.current_landuse_source.title)}
                         slug="current_landuse_source"
                         value={props.building.current_landuse_source}
                         mode={props.mode}
                         copy={props.copy}
                         onChange={props.onChange}
-                        tooltip={dataFields.current_landuse_source.tooltip}
-                        placeholder={dataFields.current_landuse_source.example}
+                        tooltip={t(dataFields.current_landuse_source.tooltip)}
+                        placeholder={t(dataFields.current_landuse_source.example)}
                         options={dataFields.current_landuse_source.items}
                         />
                     {(props.building.current_landuse_source == "Expert/personal knowledge of building" ||
                         props.building.current_landuse_source == "Online streetview image" ||
                         props.building.current_landuse_source == null) ? <></> :
                         <><MultiDataEntry
-                            title={dataFields.current_landuse_link.title}
+                            title={t(dataFields.current_landuse_link.title)}
                             slug="current_landuse_link"
                             value={props.building.current_landuse_link}
                             mode={props.mode}
                             copy={props.copy}
                             onChange={props.onChange}
-                            tooltip={dataFields.current_landuse_link.tooltip}
+                            tooltip={t(dataFields.current_landuse_link.tooltip)}
                             placeholder="https://..."
                             editableEntries={true}
                             isUrl={true}
@@ -143,14 +144,14 @@ const UseView: React.FunctionComponent<CategoryViewProps> = (props) => {
                         <div>
                             <div className={`alert alert-dark`} role="alert" style={{ fontSize: 13, backgroundColor: "#f6f8f9" }}>
                                 <i>
-                                    Below is a more general classification for the land use of this building, automatically derived from the information above.
+                                    {t('Below is a more general classification for the land use of this building, automatically derived from the information above.')}
                                 </i>
                             </div>
                         </div>
                     }
                     <DataEntry
-                        title={dataFields.current_landuse_order.title}
-                        tooltip={dataFields.current_landuse_order.tooltip}
+                        title={t(dataFields.current_landuse_order.title)}
+                        tooltip={t(dataFields.current_landuse_order.tooltip)}
                         slug="current_landuse_order"
                         value={props.building.current_landuse_order}
                         mode={props.mode}